feat: Add a standartized payload datastructure for the exchange of all client IDs that are currently connected to the server.
Add a standartized datastructure that is inserted in the payload of the message, since the transmission of
multiple elements within the payload over the network does not work with the created framework. Also I learned
that it is useful to restrict the datastructure that the client and the server can exchange in order to
get a reliable networking protocol.

No issue addressed